test(graph_slice): move parse_helpers tests to a separate module
Remove inline tests from parse_helpers.rs and move them to a new
parse_helpers_tests.rs file. This cleans up the main source file and
organizes tests into a dedicated test module. Also update mod.rs to
conditionally include the new test modules.
